JASTrack, JASNoteMgr, JASOuterParam, JASSeqParser

+++ include/JSystem/JAudio/JASSeqCtrl.h
@@ -14,7 +14,7 @@ namespace JASystem {
         bool waitCountDown();
         bool callIntr(void*);
         bool retIntr();
-        u32 get16(u32) const;
+        u16 get16(u32) const;
         u32 get24(u32) const;
         u32 get32(u32) const;
         u32 read16();
